requirements.txt
python-telegram-bot==21.6
APScheduler==3.10.4
httpx==0.27.2
aiosqlite==0.20.0
pydantic-settings==2.5.2
tenacity==9.0.0
python-dotenv==1.0.1
tzdata==2024.2

# Dev
pytest==8.3.3
pytest-asyncio==0.24.0
respx==0.21.1
